+++ frontend/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Implemented stable coin dashboard with the following sections: Overview, Supply, Holders
+- Implemented stable coin issuer specific dashboard with the following sections: Overview, Holder table, Holder distribution
+- Added environment config `enablePltFeatures` to control stable coin menu display.
+
 ## [1.7.14] - 2025-04-30
 
 ### Changed
@@ -16,13 +24,12 @@ All notable changes to this project will be documented in this file.
 - The `lastCumulativeAccountsCreated`, `lastCumulativeTransactionCount`, and `sumRewardAmount` are optional types now and set to 0 if not presen in reward metrics.
 - Extend the type `updateTransactionTypes` with the missing cases `VALIDATOR_SCORE_PARAMETERS_UPDATE`, `GAS_REWARDS_CPV_2_UPDATE`, `MINT_DISTRIBUTION_CPV_1_UPDATE`, `UPDATE_LEVEL_1_KEYS`, and `UPDATE_LEVEL_2_KEYS`.
 
+
 ## [1.7.13] - 2025-04-14
 
 ### Fixed
 
 - Support `node_statuses` id format of fungy backend when obtaining legacy backend id format in `NodeLink` component
-
-## [1.7.12] - 2025-04-10
 
 ### Fixed